- Move display related functionality into `display.h`/'display.c'.
