+++ .gitignore
@@ -10,3 +10,6 @@
 /Engine/x64/Debug
 /Engine/.idea/.idea.Engine/.idea
 /Engine/Engine.sln.DotSettings.user
+/Engine/Engine/Engine.vcxproj
+/Engine/Engine/Engine.vcxproj
+/Engine/Engine/Engine.vcxproj.filters
